backend/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# CORS and Security
python-jose[cryptography]==3.3.0

# Data Validation
pydantic==2.10.3
pydantic-settings==2.7.0

# LLM and AI
langchain==0.3.13
langchain-openai==0.2.14
langchain-anthropic==0.3.3
langchain-google-genai==2.0.5
langgraph==0.2.59
openai==1.58.1
google-generativeai==0.8.3

# Video Processing
ffmpeg-python==0.2.0

# File Operations
aiofiles==24.1.0

# Utilities
python-dotenv==1.0.1
uuid==1.30

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
